+++ core/src/sandarena/infowindow/InfoWindow.java
@@ -1,21 +1,10 @@
 package sandarena.infowindow;
 
-import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.g2d.Batch;
 import com.badlogic.gdx.scenes.scene2d.Group;
 
-import java.util.ArrayList;
-
 import sandarena.Resolution;
 import sandarena.donnee.Utili;
-import sandarena.joueur.Personnage;
-import sandarena.joueur.competence.Competence;
-import sandarena.partie.compcase.CompetenceIG;
-import sandarena.partie.compcase.PersonnageIG;
-import sandarena.partie.effet.EffetBuf;
-import sandarena.partie.gui.interfacep.empinterface.EmplacementComp;
-import sandarena.partie.gui.interfacep.empinterface.EmplacementEffet;
-import sandarena.partie.gui.interfacep.empinterface.EmplacementPerso;
 
 
 /**
@@ -24,8 +13,8 @@ import sandarena.partie.gui.interfacep.empinterface.EmplacementPerso;
 public abstract class InfoWindow extends Group {
     protected Icone icone;
     protected Info info;
-    protected static final int DIFF = 100;
-    protected static final int TAILLE_EFFET = 20;
+    protected static final byte DIFF = 100;
+    protected static final byte TAILLE_EFFET = 20;
     protected boolean affiche = false;
 
     public InfoWindow(Object emplacement) {
