+++ core/src/sandarena/joueur/competence/active/CompetenceInvocation.java
@@ -4,10 +4,11 @@ package sandarena.joueur.competence.active;
 import java.util.ArrayList;
 
 import sandarena.joueur.competence.CompetenceActive;
+import sandarena.partie.compcase.PersonnageIG;
 
 public class CompetenceInvocation extends CompetenceActive {
 
-    public CompetenceInvocation( int recharge, int utilisation, int porte, int portemin, int zone) {
+    public CompetenceInvocation( byte recharge, byte utilisation, byte porte, byte portemin, byte zone) {
         super( recharge, utilisation, porte, portemin, zone);
     }
 
@@ -15,4 +16,6 @@ public class CompetenceInvocation extends CompetenceActive {
     public ArrayList<String> toStrings() {
         return super.toStrings();
     }
+
+
 }
