+++ core/src/sandarena/joueur/competence/active/CompetenceSoin.java
@@ -4,11 +4,12 @@ package sandarena.joueur.competence.active;
 import java.util.ArrayList;
 
 import sandarena.joueur.competence.CompetenceActive;
+import sandarena.partie.compcase.PersonnageIG;
 
 
 public class CompetenceSoin extends CompetenceActive {
 
-    public CompetenceSoin( int recharge, int utilisation, int porte, int portemin, int zone) {
+    public CompetenceSoin( byte recharge, byte utilisation, byte porte, byte portemin, byte zone) {
         super( recharge, utilisation, porte, portemin, zone);
     }
 
@@ -16,4 +17,5 @@ public class CompetenceSoin extends CompetenceActive {
     public ArrayList<String> toStrings() {
         return super.toStrings();
     }
+
 }
